MB-20958 - removed disabled as field type choice
Index mapping validation in bleve doesn't allow 'disabled' as a field
type, so simplest quick fix right now is to remove disabled as a
choice in the UI drop-down.

Most of the machinery to display a 'disabled' field type in the UI,
though, is left in in case we want to actually handle this properly
and re-enable a choice of disabled.